+++ src/main/java/ru/mihaly4/vkmd/menu/MainMenu.java
@@ -3,6 +3,7 @@ package ru.mihaly4.vkmd.menu;
 import de.codecentric.centerdevice.MenuToolkit;
 import javafx.scene.control.Menu;
 import javafx.scene.control.MenuBar;
+import org.apache.commons.lang3.SystemUtils;
 import ru.mihaly4.vkmd.R;
 import ru.mihaly4.vkmd.view.AboutView;
 
@@ -17,6 +18,10 @@ public class MainMenu implements IMenu {
 
     @Override
     public void render() {
+        if (!SystemUtils.IS_OS_MAC) {
+            return;
+        }
+
         MenuBar bar = new MenuBar();
         MenuToolkit tk = MenuToolkit.toolkit();